+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [1.1.41]
+### Added
+- Add missing notification constants for row selection instrumentation
+
+### Changed
+
+### Removed
+
+
+## [1.1.40]
+### Added
+- Instrument Yahoo broker selection
+
+### Changed
+
+### Removed
+
 ## [1.1.39]
 ### Added
 - Transactions screen